I "exit_when_idle" seems to have been disabled
I commented this out because the "exit_when_idle" seems to have been disabled due to user abuse (according to various forum posts when I noticed it wasn't working) but I want to keep it here in case I am wrong.  The BOINC/SETI folks seem to understand when WU are lost, and they just reassign them since they assign the same unit to multiple people anyway, not a big deal for them.